+++ storage/nodes_test.go
@@ -3,7 +3,10 @@ package storage_test
 import (
 	"errors"
 	"sort"
+	"strconv"
 	"testing"
+
+	"github.com/stretchr/testify/require"
 
 	"code.vegaprotocol.io/data-node/logging"
 	"code.vegaprotocol.io/data-node/storage"
@@ -24,6 +27,81 @@ func (o ByXY) Less(i, j int) bool {
 	}
 
 	return o[i].Amount < o[j].Amount
+}
+
+func TestCleanupOldEpochsFromNodes(t *testing.T) {
+	nodeStore := storage.NewNode(logging.NewTestLogger(), storage.NewDefaultConfig())
+	testNode1 := pb.Node{
+		Id:               "pub_key_1",
+		PubKey:           "pub_key_1",
+		TmPubKey:         "tm_pub_key_1",
+		EthereumAdddress: "eth_pub_key_1",
+		InfoUrl:          "http://info-node-1.vega",
+		Location:         "UK",
+		Status:           pb.NodeStatus_NODE_STATUS_VALIDATOR,
+	}
+	testNode2 := pb.Node{
+		Id:               "pub_key_2",
+		PubKey:           "pub_key_2",
+		TmPubKey:         "tm_pub_key_2",
+		EthereumAdddress: "eth_pub_key_2",
+		InfoUrl:          "http://info-node-2.vega",
+		Location:         "UK",
+		Status:           pb.NodeStatus_NODE_STATUS_VALIDATOR,
+	}
+	nodeStore.AddNode(testNode1)
+	nodeStore.AddNode(testNode2)
+	for i := 0; i < 30; i++ {
+		nodeStore.AddDelegation(pb.Delegation{
+			Party:    "party1",
+			NodeId:   "pub_key_1",
+			EpochSeq: strconv.Itoa(i),
+			Amount:   "100",
+		})
+		nodeStore.AddDelegation(pb.Delegation{
+			Party:    "party1",
+			NodeId:   "pub_key_2",
+			EpochSeq: strconv.Itoa(i),
+			Amount:   "200",
+		})
+		epochSeq := strconv.Itoa(i)
+		node1, _ := nodeStore.GetByID("pub_key_1", epochSeq)
+		require.Equal(t, "100", node1.StakedByDelegates)
+
+		node2, _ := nodeStore.GetByID("pub_key_2", epochSeq)
+		require.Equal(t, "200", node2.StakedByDelegates)
+	}
+	for i := 30; i < 40; i++ {
+		nodeStore.AddDelegation(pb.Delegation{
+			Party:    "party1",
+			NodeId:   "pub_key_1",
+			EpochSeq: strconv.Itoa(i),
+			Amount:   "100",
+		})
+		nodeStore.AddDelegation(pb.Delegation{
+			Party:    "party1",
+			NodeId:   "pub_key_2",
+			EpochSeq: strconv.Itoa(i),
+			Amount:   "200",
+		})
+		// we don't have delegations for the 31st past epoch
+		epochSeqMinus30 := strconv.Itoa(i - 30)
+		node1, _ := nodeStore.GetByID("pub_key_1", epochSeqMinus30)
+		require.Equal(t, "0", node1.StakedByDelegates)
+
+		node2, _ := nodeStore.GetByID("pub_key_2", epochSeqMinus30)
+		require.Equal(t, "0", node2.StakedByDelegates)
+
+		// we have delegation for the past 30 epochs
+		for j := 0; j < 30; j++ {
+			epochSeq := strconv.Itoa(i - j)
+			node1, _ := nodeStore.GetByID("pub_key_1", epochSeq)
+			require.Equal(t, "100", node1.StakedByDelegates)
+
+			node2, _ := nodeStore.GetByID("pub_key_2", epochSeq)
+			require.Equal(t, "200", node2.StakedByDelegates)
+		}
+	}
 }
 
 func TestNodes(t *testing.T) {